Replace LoggerT with LoggerEffect
